+++ tests/integration/conftest.py
@@ -177,13 +177,22 @@ def batch_project(client, rand_gen):
 
 @pytest.fixture
 def consensus_project(client, rand_gen):
-    project = client.create_project(name=rand_gen(str), auto_audit_percentage=0)
+    project = client.create_project(name=rand_gen(str),
+                                    auto_audit_percentage=0,
+                                    queue_mode=QueueMode.Dataset)
     yield project
     project.delete()
 
 
 @pytest.fixture
 def dataset(client, rand_gen):
+    dataset = client.create_dataset(name=rand_gen(str))
+    yield dataset
+    dataset.delete()
+
+
+@pytest.fixture(scope='function')
+def unique_dataset(client, rand_gen):
     dataset = client.create_dataset(name=rand_gen(str))
     yield dataset
     dataset.delete()